Delete Tenant Get Rental Listing Create Rental Listing

+++ Tables/RentalListing.cs
@@ -5,7 +5,7 @@
         public int ListingID { get; set; }
         public string? PropertyName { get; set; }
         public string? Address { get; set; }
-        public int? MonthlyRent { get; set; }
+        public decimal? MonthlyRent { get; set; }
         public DateOnly? AvailableDate { get; set; }
     }
 }
